requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.26.0
python-multipart==0.0.9
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
python-dotenv==1.0.1
